// cosmowcl/src/app/shared/models/api/cosmoauthorization/tipoSegnalibro.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * preferenze-utente
 * API per la gestione delle preferenze dell\'utente
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TipoSegnalibro { 
    codice: string;
    descrizione?: string;
}